requirements.txt
pymongo[srv]==4.7.2
python-dotenv==1.0.1
bcrypt==4.1.3
python-dateutil
anthropic
groq==1.2.0
streamlit
streamlit-autorefresh==1.0.1